+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_03_24_123000) do
+ActiveRecord::Schema[8.1].define(version: 2026_07_06_140000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -47,6 +47,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_03_24_123000) do
     t.date "month", null: false
     t.text "notes"
     t.datetime "updated_at", null: false
+    t.index ["category_id", "month"], name: "index_budgets_on_category_id_and_month", unique: true
     t.index ["category_id"], name: "index_budgets_on_category_id"
   end